+++ utils/Functions.js
@@ -43,67 +43,66 @@ var Functions;
     /** Create a function that lazily returns a computed value
      * @param initializer: the function that initializes the lazy field and returns it (this function will only be called once)
      * @return a function that returns the cached value returned by the {@code initializer} function
+     * with an optiona 'refetch' parameter which, if true, forces the initializer to be called again
      */
     function lazyField(initializer) {
         var value = null;
-        return function lazyInitializer() {
-            if (value == null) {
+        return function lazyInitializer(refetch) {
+            if (value == null || refetch === true) {
                 value = initializer();
             }
             return value;
         };
     }
     Functions.lazyField = lazyField;
-    /** Create a function that takes 1 argument and lazily returns a computed value
-     * @see createLazyInitializer()
-     */
-    function lazyGetter1Arg(initializer) {
-        var value = null;
-        return function lazyInitializer(arg) {
-            if (value == null) {
-                value = initializer(arg);
-            }
-            return value;
-        };
+    function partial(func) {
+        var args = [];
+        for (var _i = 1; _i < arguments.length; _i++) {
+            args[_i - 1] = arguments[_i];
+        }
+        switch (func.length) {
+            case 1:
+                switch (args.length) {
+                    case 0: return function partial1Bind0(a) { return func(a); };
+                    case 1: return function partial1Bind1() { return func(args[0]); };
+                    default: return function partial1BindMany() { return func.apply(undefined, arguments); };
+                }
+            case 2:
+                switch (args.length) {
+                    case 0: return function partial2Bind0(a, b) { return func(a, b); };
+                    case 1: return function partial2Bind1(b) { return func(args[0], b); };
+                    case 2: return function partial2Bind2() { return func(args[0], args[1]); };
+                    default: return function partial2BindMany() { return func.apply(undefined, arguments); };
+                }
+            case 3:
+                switch (args.length) {
+                    case 0: return function partial3Bind0(a, b, c) { return func(a, b, c); };
+                    case 1: return function partial3Bind1(b, c) { return func(args[0], b, c); };
+                    case 2: return function partial3Bind2(c) { return func(args[0], args[1], c); };
+                    case 3: return function partial3Bind3() { return func(args[0], args[1], args[2]); };
+                    default: return function partial3BindMany() { return func.apply(undefined, arguments); };
+                }
+            default:
+                switch (args.length) {
+                    case 0:
+                        return function partialManyBindNone() {
+                            return func.apply(undefined, arguments);
+                        };
+                    default:
+                        return function partialManyBindMany() {
+                            var allArgs = new Array(args.length);
+                            for (var i = 0, s = args.length; i < s; i++) {
+                                allArgs[i] = args[i];
+                            }
+                            for (var i = 0, s = arguments.length; i < s; i++) {
+                                allArgs.push(arguments[i]);
+                            }
+                            return func.apply(undefined, allArgs);
+                        };
+                }
+        }
     }
-    Functions.lazyGetter1Arg = lazyGetter1Arg;
-    /** Create a function that takes 2 arguments and lazily returns a computed value
-     * @see createLazyInitializer()
-     */
-    function lazyGetter2Arg(initializer) {
-        var value = null;
-        return function lazyInitializer(arg1, arg2) {
-            if (value == null) {
-                value = initializer(arg1, arg2);
-            }
-            return value;
-        };
-    }
-    Functions.lazyGetter2Arg = lazyGetter2Arg;
-    /** Create a function that calls another function with 1 pre-specified argument and returns the result
-     */
-    function wrap1Arg(func, arg1) {
-        return function funcWrapper1() {
-            return func(arg1);
-        };
-    }
-    Functions.wrap1Arg = wrap1Arg;
-    /** Create a function that calls another function with 2 pre-specified arguments and returns the result
-     */
-    function wrap2Arg(func, arg1, arg2) {
-        return function funcWrapper2() {
-            return func(arg1, arg2);
-        };
-    }
-    Functions.wrap2Arg = wrap2Arg;
-    /** Create a function that calls another function with 3 pre-specified arguments and returns the result
-     */
-    function wrap3Arg(func, arg1, arg2, arg3) {
-        return function funcWrapper3() {
-            return func(arg1, arg2, arg3);
-        };
-    }
-    Functions.wrap3Arg = wrap3Arg;
+    Functions.partial = partial;
     /** Creates a new function that wraps a given function.
      * Useful for logging performance, number of calls, etc.
      * @param func: the function to call each time this function is called